reorganize into submodules
